refactor: Moved logo components into a new file
